transactions-generator/requirements.txt
kafka-python==2.0.2
numpy
python-dotenv